% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_denv}
\alias{theme_denv}
\title{Denv's ggplot2 theme}
\usage{
theme_denv(font = "Lato", base.size = 12, text.color = "gray20",
  grid.color = "gray80", background.color = "papayawhip",
  legend.pos = "bottom")
}
\arguments{
\item{font}{Choose base font for theme, requires library(extrafont) - defauls to "Lato"}

\item{base.size}{Base size for text}
}
\description{
This function allows you to express your love of cats.
}
\examples{
theme_denv()
}
